Consolidate benchmarks: shared MP interface, deduplicate code
- benchmarks-util: multiplatform module with ProtobufBenchmarkSet
  interface (kotlinx-benchmark Blackhole), randomUtf8String (kotlin
  Random), readDatasetBytes expect/actual
- JvmProtobufBenchmarkSet: extends base with InputStream-streaming
  methods for protobuf-java and Wire
- Remove duplicate ProtoktBenchmarks.kt from jvmMain (commonMain
  version covers all benchmarks)
- protobuf-java/Wire benchmarks: switch to kotlin.random.Random,
  implement JvmProtobufBenchmarkSet
