+++ benchmarks/benchmarks-util/build.gradle.kts
@@ -14,5 +14,21 @@
  */
 
 plugins {
-    id("protokt.benchmarks-conventions")
+    id("protokt.multiplatform-conventions")
+}
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(libs.kotlinx.benchmark.runtime)
+            }
+        }
+
+        val jvmMain by getting {
+            dependencies {
+                implementation(libs.jmh.core)
+            }
+        }
+    }
 }
