+++ benchmarks/benchmarks-util/src/commonMain/kotlin/protokt/v1/benchmarks/RandomUtf8.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2026 Toast, Inc.
+ * Copyright (c) 2019 Toast, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
