+++ benchmarks/wire-benchmarks/src/main/kotlin/protokt/v1/benchmarks/WireBenchmarks.kt
@@ -35,13 +35,13 @@ import org.openjdk.jmh.annotations.State
 import org.openjdk.jmh.infra.Blackhole
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
-import java.util.Random
+import kotlin.random.Random
 import java.util.concurrent.TimeUnit
 
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
-open class WireBenchmarks : ProtobufBenchmarkSet {
+open class WireBenchmarks : JvmProtobufBenchmarkSet {
     private lateinit var largeDataset: BenchmarkDataset
     private lateinit var largeParsedDataset: List<GenericMessage1>
     private lateinit var mediumDataset: BenchmarkDataset
